+++ game/src/Ship.h
@@ -2,6 +2,8 @@
 #include <iostream>
 #include <utility>
 #include <initializer_list>
+#include <cstdlib>
+#include <ctime>
 #include <list>
 #include <cmath>
 
@@ -13,21 +15,21 @@ class Ship{
     };
     std::vector<segmentState> segments;
     bool is_vertical = true;
-    std::pair<size_t, size_t> coordinates;
-    size_t length;
+    std::pair<int, int> coordinates;
+    int length;
     public:
         Ship() = default;
-        Ship(size_t length, std::pair<size_t, size_t> coordinates, bool is_vertical);
+        Ship(int length, std::pair<int, int> coordinates, bool is_vertical);
         Ship(const Ship &ship);
         Ship& operator = (const Ship& ship);
         Ship(Ship && ship) noexcept;
         Ship& operator = (Ship && ship) noexcept;
 
-        size_t getLen() const;
-        std::pair<size_t, size_t> getCoor() const;
+        int getLen() const;
+        std::pair<int, int> getCoor() const;
         bool IsVertical() const;
-        segmentState getSegment(size_t index) const;
+        segmentState getSegment(int index) const;
         
-        void Attack(size_t index);
+        void Attack(int index);
         bool isDestroyed();
 };
